Move cardinality checker and add stub test file
