+++ tests/rule_based_profiler/domain_builder/test_cardinality_checker.py
@@ -1,0 +1,10 @@
+def test_cardinality_checker_instantiation_valid_parameters():
+    raise NotImplementedError
+
+
+def test_cardinality_checker_instantiation_invalid_parameters():
+    raise NotImplementedError
+
+
+def test_exhaustively_all_supported_cardinality_limits():
+    raise NotImplementedError
